fix(lazyvim-telescope): the `config` function of `telescope.nvim` in undotree.lua was not called
If you have multiple specs for the same plugin, then all `opts` will be evaluated, but only the last `config`.
So `config` function in undotree.lua and zoxide.lua can not be called both.
Fix it by using `opts` instead of `config` for multiple specs of same `telescope.nvim` plugin.